requirements.txt
discord.py
aiosqlite
python-dotenv
flask
vk_api
scikit-learn
matplotlib
gunicorn